feat(UserProfilePage):add user about, grouped_info and note widget -UL-97
