+++ lib/contacts/blocked_page.dart
@@ -118,6 +118,7 @@ class BlockedPage extends StatelessWidget {
                     friendNum: 24,
                     relationship: Relationship.block,
                     friendRequestSent: false,
+                    // isBlocked: true,
                     about:
                         'Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae beatae vitae dicta sunt explicabo. ',
                   );
