+++ lib/contacts/user_profile_page/models/models_export.dart
@@ -1,5 +1,7 @@
 export 'app_bar_with_banner.dart';
 export 'cta_button.dart';
-export 'network_profiles_body.part.dart';
-export 'profile_data_body.part.dart';
+export 'user_about.dart';
 export 'user_basic_info.dart';
+export 'user_grouped_info.dart';
+export 'user_linked_accounts.dart';
+export 'user_note.dart';
